apps/backend/requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# CORS and middleware
python-dotenv>=1.0.0

# WebSocket support
websockets>=12.0

# HTTP client
httpx>=0.25.0
aiohttp>=3.9.0

# Basic utilities
python-dateutil>=2.8.0
pytz>=2023.3